stopped propagation of edit and delete modals moved delete modal to modals folder
